Add admin permissions management functionality. Introduce handlers for editing admin permissions, including permission selection and toggling. Update keyboards and language dictionaries to support new features. Enhance access control with permission filters for various admin actions.

+++ admin/ban/handlers.py
@@ -14,7 +14,8 @@ from telegram.ext import (
     MessageHandler,
     filters,
 )
-from custom_filters import PrivateChatAndAdmin
+from custom_filters import PrivateChatAndAdmin, PermissionFilter
+from models import Permission
 from common.keyboards import (
     build_admin_keyboard,
     build_back_button,
@@ -29,7 +30,7 @@ USER, CONFIRM = range(2)
 
 
 async def ban_unban(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    if PrivateChatAndAdmin().filter(update):
+    if PrivateChatAndAdmin().filter(update) and PermissionFilter(Permission.BAN_USERS).filter(update):
         lang = get_lang(update.effective_user.id)
         await update.callback_query.delete_message()
         await context.bot.send_message(
@@ -53,7 +54,7 @@ async def ban_unban(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
 
 async def get_user_id(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    if PrivateChatAndAdmin().filter(update):
+    if PrivateChatAndAdmin().filter(update) and PermissionFilter(Permission.BAN_USERS).filter(update):
         lang = get_lang(update.effective_user.id)
         if update.effective_message.users_shared:
             user_id = update.effective_message.users_shared.users[0].user_id
@@ -115,7 +116,7 @@ async def get_user_id(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
 
 async def confirm_ban_unban(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    if PrivateChatAndAdmin().filter(update):
+    if PrivateChatAndAdmin().filter(update) and PermissionFilter(Permission.BAN_USERS).filter(update):
         lang = get_lang(update.effective_user.id)
         user_id = context.user_data["user_id_to_ban_unban"]
         with models.session_scope() as s:
